services/ingest-worker/requirements.txt
watchdog==4.0.1
pillow==10.4.0
requests==2.32.3
pypdf==5.0.1
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
lancedb==0.7.3
pyarrow==17.0.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
